Factory-ize wrapper creation to allow React Native support
