+++ .gitignore
@@ -3,3 +3,4 @@ examples/**/bundle.js
 node_modules
 coverage
 npm-debug.log
+.idea/
