+++ README.md
@@ -205,6 +205,20 @@ const getRoutes = (store) => {
 };
 ```
 
+### React Native
+Usage as above except include the react-native implementation
+
+
+```js
+import { UserAuthWrapper } from 'redux-auth-wrapper/native';
+
+const UserIsAuthenticated = UserAuthWrapper({
+  authSelector: state => state.user,
+  redirectAction: routerActions.replace,
+  wrapperDisplayName: 'UserIsAuthenticated'
+})
+```
+
 ### Other examples
 * [Redux-Router and React-Router 1.0 with JWT](https://github.com/mjrussell/react-redux-jwt-auth-example/tree/auth-wrapper)
 * [React-Router-Redux and React-Router 2.0 with JWT](https://github.com/mjrussell/react-redux-jwt-auth-example/tree/react-router-redux)
