+++ package.json
@@ -10,7 +10,7 @@
     "src"
   ],
   "scripts": {
-    "build": "mkdir -p lib && babel ./src/index.js --out-file ./lib/index.js",
+    "build": "mkdir -p lib && babel ./src --out-dir ./lib",
     "lint": "eslint src test",
     "prepublish": "rm -rf lib && npm run build",
     "test": "mocha --compilers js:babel-core/register --recursive --require test/init.js test/**/*-test.js",
@@ -45,6 +45,7 @@
     "mocha": "^2.3.4",
     "react": "~0.14.3",
     "react-addons-test-utils": "^0.14.6",
+    "react-native": "^0.24.1",
     "react-redux": "^4.0.1",
     "react-router": "2.0.1",
     "react-router-redux": "~4.0.0",
