+++ src/native.js
@@ -1,0 +1,4 @@
+import React, { View } from 'react-native';
+import factory from './factory';
+
+export const UserAuthWrapper = factory(React, View);
